requirements.txt
langgraph
langchain-openai
langchain
feedparser
arxiv
beautifulsoup4
python-dotenv
requests
slack-sdk
pydantic
pytz
python-dateutil
jinja2
